Rename Module -> Behavior, Update Documentation
