+++ README.YARD.md
@@ -8,5 +8,5 @@ This is a design choice. It does create some pressure on
 garbage collector, but completely eliminates all concurrency
 and aliasing problems.
 
-See also {AST::Node}, {AST::Processor} and {AST::Sexp} for additional
-recommendations and design patterns.
+See also {AST::Node}, {AST::Processor::Behavior} and {AST::Sexp} for
+additional recommendations and design patterns.
